third_party/WebKit/LayoutTests/bluetooth/getCharacteristic/blacklisted-characteristic.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('BlacklistTestAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['device_information']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('device_information'))
    .then(service => {
      return assert_promise_rejects_with_message(
        service.getCharacteristic('serial_number_string'),
        new DOMException('getCharacteristic(s) called with blacklisted UUID. ' +
                         'https://goo.gl/4NeimX',
                         'SecurityError'),
        'Serial Number String characteristic is blacklisted.');
    });
}, 'Serial Number String characteristic is blacklisted.');
</script>
